+++ e2e/tests/manageCases/caseProgression/checkApplication/checkApplication.spec.ts
@@ -3,7 +3,7 @@ import Config from "../../../../config";
 import createDaCitizenCourtNavCase from "../../../../common/createCaseHelper";
 import { Helpers } from "../../../../common/helpers";
 import config from "../../../../config";
-import { CheckApplicationJourney } from "../../../../journeys/manageCases/caseProgression/e2eFlowUpToServiceOfApplication/checkApplication/checkApplicationJourney.ts";
+import { CheckApplicationJourney } from "../../../../journeys/manageCases/caseProgression/checkApplication/checkApplicationJourney.ts";
 
 test.use({ storageState: Config.sessionStoragePath + "caseWorker.json" });
